+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.0.1
+## 1.1.0
 
 * Add V2 Clients for Text tracks, Tags, Usage, Player Bidding Configs, Originals, Media Renditions, Thumbnails.
 * Update Playlists client to include watchlist.